+++ include/highfive/H5Reference.hpp
@@ -53,7 +53,6 @@ class Reference {
     /// \return the ObjectType of the referenced object
     ObjectType getType(const Object& location) const;
 
-  protected:
     /// \brief Create a Reference from a low-level HDF5 object reference
     inline explicit Reference(const hobj_ref_t h5_ref)
         : href(h5_ref){};
@@ -64,6 +63,7 @@ class Reference {
     /// be stored
     void create_ref(hobj_ref_t* refptr) const;
 
+  protected:
   private:
     Object get_ref(const Object& location) const;
